variable change

+++ python_variable_annotations/2-floor.py
@@ -8,7 +8,7 @@ and returns the floor of the float
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """
     Returns the floor of the float
     """
